requirements.txt
fastapi==0.116.0
pydantic==2.11.7
python-dotenv==1.1.1
uvicorn==0.35.0